compose: Add types to `withSafeTimeout`
